refactor: move string util to string util file
