+++ src/shared/util/string.ts
@@ -48,3 +48,22 @@ export const ellipsify = (input: string, chars: number): string => {
     }
     return ellipisifed;
 };
+
+/**
+ *  Stringifies a list of items in English format.
+ *
+ * @param items - The list of items to stringify.
+ * @returns A string representation of the list in English format.
+ * @example
+ * englishStringifyList([]) // ''
+ * englishStringifyList(['Alice']) // 'Alice'
+ * englishStringifyList(['Alice', 'Bob']) // 'Alice and Bob'
+ * englishStringifyList(['Alice', 'Bob', 'Charlie']) // 'Alice, Bob, and Charlie'
+ */
+export const englishStringifyList = (items: readonly string[]): string => {
+    if (items.length === 0) return '';
+    if (items.length === 1) return items[0]!;
+    if (items.length === 2) return `${items[0]} and ${items[1]}`;
+
+    return `${items.slice(0, -1).join(', ')}, and ${items.at(-1)}`;
+};
